# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adeptus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/engagements/{engagement_id}/chat/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Chat Messages",
        "description": "List the calling user's private chat messages for an engagement, oldest-first.\n\nReturns only the caller's own conversation (\u00a75.4 per-user privacy). Requires\nmembership; a non-member or missing engagement returns 404 (no existence disclosure).",
        "operationId": "list_chat_messages",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessagePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Chat Message",
        "description": "Persist a user message and an empty pending assistant message, then return both.\n\nStream the assistant reply over ``WS /ws/chat/{assistant_message_id}``. Requires\nmembership (404). The two 409 cases are translated inline to the EgressConfirmationRequired\nbody (Slice 14): an archived engagement (\u00a74) \u2192 ``engagement_archived``; a cloud-enabled\nsecret-matching send without ``confirmed_egress`` \u2192 ``egress_secret_flagged`` with the\nmatched category NAMES (never the secret value, \u00a75.5).",
        "operationId": "send_chat_message",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendChatMessageResult"
                }
              }
            }
          },
          "409": {
            "description": "Either the engagement is archived (read-only, \u00a74), OR the engagement is cloud-enabled and the message matched a likely-secret pattern but was not confirmed (\u00a75.1 pattern-friction). The body's reason distinguishes the two; clients re-send with confirmed_egress=true to proceed past the friction case.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EgressConfirmationRequired"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engagements/{engagement_id}/chat/messages/{message_id}/debug": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Turn Debug",
        "description": "Return the AI debug record (\u00a714) for one of the caller's own assistant turns.\n\nThe exact \u00a75.3 relevant subset of the graph injected, the raw prompt, and the model\noutput. Membership + ownership scoped: a non-member, non-owner, wrong-engagement, or\nnon-assistant message all return 404 (no existence disclosure, \u00a717.1 / \u00a75.4).",
        "operationId": "get_chat_turn_debug",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnDebug"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Message not found, not owned by caller, or not an assistant turn"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 32768,
            "minLength": 1,
            "title": "Content",
            "description": "The user's message text, sent verbatim to the model (no redaction, \u00a75.5)."
          },
          "pinned_node_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Pinned Node Ids",
            "description": "Node ids the user has pinned (client-side pinStore, \u00a75.4). Always-included arm of the \u00a75.3 union. Unknown/foreign ids are ignored server-side (\u00a717.1)."
          },
          "recent_node_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Recent Node Ids",
            "description": "Node ids recently interacted with this session, most-recent-first; the \"last N nodes touched in the conversation\" arm (\u00a75.3). Server truncates to N."
          },
          "mentioned_node_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Mentioned Node Ids",
            "description": "Node ids @-mentioned in recent messages (\u00a75.3). Empty until the @-mention UI (Slice 31); accepted now for forward-compatibility. Server truncates to K."
          },
          "confirmed_egress": {
            "type": "boolean",
            "title": "Confirmed Egress",
            "description": "The user has seen the cloud egress-friction modal for THIS content and chose to send it unmodified anyway (\u00a75.1, Slice 14). Consulted only when the engagement is cloud_enabled and the content matched a secret pattern; ignored otherwise. Never suppresses the audit record (a confirmed send is audited as confirmed).",
            "default": false
          },
          "persona_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Id",
            "description": "The persona whose system prompt should shape THIS turn (\u00a75.3, Slice 15). Must be a built-in or one of the caller's own personas; an unknown/foreign id falls back to the `general` built-in server-side (\u00a717.1, never errors). Null/absent \u2192 `general`. Chosen per send so the user can switch persona mid-chat without resetting the conversation."
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageCreate",
        "description": "Request body for POST .../chat/messages.\n\nThe three id lists carry the client-supplied arms of the \u00a75.3 \"relevant subset\"\nunion: ``pinned_node_ids`` (the always-included pinned set, \u00a75.4), ``recent_node_ids``\n(the \"last N touched in the conversation\" approximation, most-recent-first), and\n``mentioned_node_ids`` (the @-mention arm, empty until Slice 31). They are *inputs* only\n\u2014 the server re-resolves each id against the engagement's live graph (foreign/unknown\nids are ignored, \u00a717.1) and runs the keyword arm itself, so the client can never inject\nnode content, only ids."
      },
      "ChatMessagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageRead"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "low_confidence_threshold": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Low Confidence Threshold",
            "description": "Certainty %% below which a claim renders as low-confidence (\u00a75.3). This is the single backend tunable (ADEPTUS_CHAT_LOW_CONFIDENCE_THRESHOLD) surfaced to the UI so the frontend reads one source of truth, not a hard-coded mirror.",
            "default": 70
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ChatMessagePage",
        "description": "A page of chat messages (oldest-first) with an opaque cursor for the next\n(older) page; ``next_cursor`` is null on the last page."
      },
      "ChatMessageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "engagement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Engagement Id"
          },
          "role": {
            "$ref": "#/components/schemas/ChatRole"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "status": {
            "$ref": "#/components/schemas/ChatMessageStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/PlanStep"
            },
            "type": "array",
            "title": "Plan",
            "description": "The AI's running plan as of this turn (\u00a75.3). Empty for user/pending/pre-slice rows; populated from the assistant row's parsed metadata so a reloaded conversation re-renders the Plan panel without the debug call."
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims",
            "description": "Certainty-tagged claims parsed from this turn (\u00a75.3). Empty when none; drives the inline certainty badges and the Graph-pane overlay."
          },
          "persona_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Id",
            "description": "The persona used for this assistant turn (\u00a75.3, Slice 15). Null for user/pending/pre-slice rows. A soft reference: the persona may since have been renamed/deleted."
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name",
            "description": "The persona's display name at turn time, denormalized onto the turn so a renamed/deleted persona still labels the historical turn (the in-chat persona chip). Null for user/pending/pre-slice rows."
          }
        },
        "type": "object",
        "required": [
          "id",
          "engagement_id",
          "role",
          "content",
          "status",
          "created_at"
        ],
        "title": "ChatMessageRead",
        "description": "One chat message as exposed by the read/write API."
      },
      "ChatMessageStatus": {
        "type": "string",
        "enum": [
          "complete",
          "pending",
          "failed"
        ],
        "title": "ChatMessageStatus",
        "description": "Lifecycle of a chat message. Mirrors ``models.CHAT_STATUSES``.\n\n``user`` messages are always ``complete``. An ``assistant`` message is ``pending``\nuntil its stream finishes (``complete``) or errors (``failed``)."
      },
      "ChatRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant"
        ],
        "title": "ChatRole",
        "description": "Chat message author role. Mirrors ``models.CHAT_ROLES``."
      },
      "ChatTurnDebug": {
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Message Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "status": {
            "$ref": "#/components/schemas/ChatMessageStatus"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphSubsetNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphSubsetEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "context_block": {
            "type": "string",
            "title": "Context Block",
            "description": "The exact graph-context text prepended to the system prompt this turn."
          },
          "raw_prompt": {
            "type": "string",
            "title": "Raw Prompt",
            "description": "The full prompt sent to the model (\u00a714 \"raw prompts\")."
          },
          "model_output": {
            "type": "string",
            "title": "Model Output",
            "description": "The model's raw output for this turn, INCLUDING the structured <adeptus-meta> block (\u00a714 \u2014 the debug view shows the unstripped output so a power user can see exactly what was parsed). Empty while pending/failed."
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/PlanStep"
            },
            "type": "array",
            "title": "Plan",
            "description": "The plan parsed from this turn's metadata block (\u00a75.3 / \u00a714)."
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims",
            "description": "The certainty claims parsed from this turn's metadata block (\u00a75.3 / \u00a714)."
          },
          "persona_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Id",
            "description": "The persona that shaped this turn (\u00a75.3 / \u00a717.6, Slice 15). Null for user/pending/pre-slice rows. The raw_prompt's leading system content is this persona's prompt, so the panel shows exactly which persona was used."
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name",
            "description": "The persona's display name at turn time (\u00a75.3). Null when no persona."
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "model",
          "status",
          "nodes",
          "edges",
          "context_block",
          "raw_prompt",
          "model_output"
        ],
        "title": "ChatTurnDebug",
        "description": "The AI debug record for one assistant turn (\u00a714): the exact relevant subset of the\ngraph injected, the raw prompt, and the model output.\n\nRead by ``GET .../chat/messages/{message_id}/debug``. ``model_output`` is the assistant\nrow's ``content`` (empty while pending/failed). ``nodes``/``edges``/``context_block``/\n``raw_prompt`` are read back from the persisted ``graph_context`` JSONB; an assistant row\nthat completed before this slice (or an empty-graph turn) has an empty subset and an\nempty ``context_block``."
      },
      "Claim": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The claim the AI flagged, verbatim (\u00a75.5)."
          },
          "certainty": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Certainty",
            "description": "Stated certainty percentage for this claim (\u00a75.3)."
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id",
            "description": "The graph node this claim is about, if any; validated against the engagement's live graph (foreign/unknown ids dropped, \u00a717.1). Drives the Graph-pane badge."
          }
        },
        "type": "object",
        "required": [
          "text",
          "certainty"
        ],
        "title": "Claim",
        "description": "One AI claim flagged with a stated certainty percentage (\u00a75.3 uncertainty signaling).\n\n``certainty`` is bounded 0\u2013100 (the parser clamps out-of-range values before this model\nis constructed). ``node_id`` is the graph node the claim is about, if any \u2014 validated at\nfinalize against the engagement's live graph (foreign/unknown ids dropped, \u00a717.1) so the\nGraph-pane certainty badge never points at a foreign/hallucinated node. ``text`` is\nrendered verbatim (no redaction, \u00a75.5)."
      },
...
      "EgressConfirmationRequired": {
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/EgressRefusalReason",
            "description": "Why the POST was refused with 409."
          },
          "matched_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Categories",
            "description": "Names of the secret-pattern categories the content matched (\u00a75.1). Empty for the archived reason. NEVER contains the matched value (\u00a75.5) \u2014 only the category name."
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "EgressConfirmationRequired",
        "description": "Body of the POST .../chat/messages 409 (Slice 14, \u00a75.1).\n\n``matched_categories`` are pattern category NAMES only (e.g. ``\"aws_access_key\"``) \u2014 NEVER\nthe matched secret value (\u00a75.5 / Risk 7) \u2014 for the friction modal's copy. Empty for the\n``engagement_archived`` reason. A client re-sends with ``confirmed_egress=true`` to proceed\npast the friction case."
      },
      "EgressRefusalReason": {
        "type": "string",
        "enum": [
          "egress_secret_flagged",
          "engagement_archived"
        ],
        "title": "EgressRefusalReason",
        "description": "Why a POST .../chat/messages was refused with 409 (Slice 14).\n\nThe single POST 409 covers two cases with a distinguishable body: a cloud-enabled send\nthat matched a likely-secret pattern without confirmation (\u00a75.1 pattern-friction), and an\narchived engagement (\u00a74 read-only). The reason lets the client tell them apart."
      },
...
      "GraphSubsetEdge": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "relation": {
            "type": "string",
            "title": "Relation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "target_id",
          "relation"
        ],
        "title": "GraphSubsetEdge",
        "description": "One edge between two selected nodes (only edges with both endpoints in the subset)."
      },
      "GraphSubsetNode": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/GraphSubsetReason"
            },
            "type": "array",
            "title": "Reasons",
            "description": "Why this node was included (one node may have several reasons)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "label",
          "reasons"
        ],
        "title": "GraphSubsetNode",
        "description": "One graph node included in the per-turn subset, tagged with its inclusion reason(s).\n\nLabels/properties are NOT carried here \u2014 only the rendered ``context_block`` (on\n``ChatTurnDebug``) holds the verbatim text sent to the model (\u00a75.5). This row is the\ngrouping/affordance data the debug panel renders (id + type + label + reasons)."
      },
      "GraphSubsetReason": {
        "type": "string",
        "enum": [
          "pinned",
          "recent",
          "mentioned",
          "keyword"
        ],
        "title": "GraphSubsetReason",
        "description": "Why a node entered the \u00a75.3 relevant subset. A node may carry several reasons."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanStep": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step",
            "description": "The todo-list item text, verbatim (\u00a75.5)."
          },
          "status": {
            "$ref": "#/components/schemas/PlanStepStatus",
            "description": "The step's current state (todo/in_progress/done)."
          }
        },
        "type": "object",
        "required": [
          "step",
          "status"
        ],
        "title": "PlanStep",
        "description": "One ordered todo-list item the AI is tracking this turn (\u00a75.3 visible plan).\n\nRendered verbatim in the inline Plan panel (no redaction, \u00a75.5). Parsed server-side\nfrom the model's trailing ``<adeptus-meta>`` block; an absent/invalid ``status`` is\ncoerced to ``todo`` by the parser before this model is constructed (task 2)."
      },
...
      "SendChatMessageResult": {
        "properties": {
          "user_message": {
            "$ref": "#/components/schemas/ChatMessageRead"
          },
          "assistant_message": {
            "$ref": "#/components/schemas/ChatMessageRead"
          }
        },
        "type": "object",
        "required": [
          "user_message",
          "assistant_message"
        ],
        "title": "SendChatMessageResult",
        "description": "Result of POST .../chat/messages: the persisted user message plus a ``pending``\nassistant placeholder. Stream the assistant reply via WS using its id."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}